Move Completion event to Events directory
