+++ src/ParallelAlgorithms/Events/Completion.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "ParallelAlgorithms/EventsAndTriggers/Completion.hpp"
+#include "ParallelAlgorithms/Events/Completion.hpp"
 
 namespace Events {
 PUP::able::PUP_ID Completion::my_PUP_ID = 0;  // NOLINT
