+++ tests/Unit/ParallelAlgorithms/EventsAndTriggers/CMakeLists.txt
@@ -4,7 +4,6 @@
 set(LIBRARY "Test_EventsAndTriggers")
 
 set(LIBRARY_SOURCES
-  Test_Completion.cpp
   Test_EventsAndTriggers.cpp
   Test_RunEventsOnFailure.cpp
   Test_Tags.cpp
@@ -17,6 +16,7 @@ target_link_libraries(
   ${LIBRARY}
   PRIVATE
   DataStructures
+  Events
   EventsAndTriggers
   ErrorHandling
   Parallel
